Smaller Changes Than Before:
Added a Port number to the header of the data (in PacketHelp)
Added a getPort to acquire this port (in PacketHelp)
Added a little bit of code for receiver. (Still incomplete and needs
some work. I think I've figured out a cleaner way to do it, but I've
decided to code this idea when my mind is fresher.)